+++ apps/admin/src/components/users/details/user-sensors-card.tsx
